+++ frontend/src/App.vue
@@ -1,6 +1,7 @@
 <template>
   <v-app>
     <v-main>
+        <Login></Login>
         <router-view/>
     </v-main>
   </v-app>
@@ -9,6 +10,7 @@
 <script>
 import { RouterView } from 'vue-router';
 import Home from './components/Home.vue';
+import Login from './components/Login.vue';
 import Search from './components/Search.vue';
 import SearchResult from './components/SearchResult.vue';
 import SearchHistory from './components/SearchHistory.vue';
@@ -24,6 +26,7 @@ export default {
     SearchResult,
     Search,
     Home,
+    Login,
     SearchHistory,
     SearchFavorite,
   },
