+++ frontend/src/assets/css/search.css
@@ -330,7 +330,7 @@ font-size: large;
 color: #f8f8f8;
 background-color: gray;
 border-radius: 10px;
-padding:7px 50px;
+padding:7px 30px;
 }
 
 #register-button:hover {
